requirements.txt
customtkinter
Pillow
torch
transformers
diffusers